leader_final_site/requirements.txt
asgiref==3.3.1
awesome-slugify==1.6.5
Django==3.1.3
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.11.0
jsonfield==3.1.0
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.4
regex==2020.11.13
six==1.15.0
sqlparse==0.4.1
Unidecode==0.4.21
